+++ demo/pages/f7material/forms-selects.vue
@@ -10,8 +10,7 @@
       <div class="content-block">Framework7 allows you to easily convert your usual form selects to dynamic pages with radios:</div>
       <div class="list-block">
         <ul>
-          <li>
-            <a href="#" class="item-link smart-select">
+          <li><a href="#" class="item-link smart-select">
               <select name="fruits">
                 <option value="apple" selected="selected">Apple</option>
                 <option value="pineapple">Pineapple</option>
@@ -25,11 +24,8 @@
                 <div class="item-inner">
                   <div class="item-title">Fruit</div>
                 </div>
-              </div>
-            </a>
-          </li>
-          <li>
-            <a href="#" class="item-link smart-select">
+              </div></a></li>
+          <li><a href="#" class="item-link smart-select">
               <select name="car" multiple="multiple">
                 <optgroup label="Japanese">
                   <option value="honda" selected="selected">Honda</option>
@@ -56,11 +52,8 @@
                 <div class="item-inner">
                   <div class="item-title">Car</div>
                 </div>
-              </div>
-            </a>
-          </li>
-          <li>
-            <a href="#" class="item-link smart-select">
+              </div></a></li>
+          <li><a href="#" class="item-link smart-select">
               <select name="mac-windows">
                 <option value="mac" selected="selected">Mac</option>
                 <option value="windows">Windows</option>
@@ -69,9 +62,7 @@
                 <div class="item-inner">
                   <div class="item-title">Mac or Windows</div>
                 </div>
-              </div>
-            </a>
-          </li>
+              </div></a></li>
         </ul>
       </div>
     </div>